move builder into stores
