+++ tools/goctl/model/sql/template/import.go
@@ -12,7 +12,7 @@ var (
 	"github.com/tal-tech/go-zero/core/stores/sqlc"
 	"github.com/tal-tech/go-zero/core/stores/sqlx"
 	"github.com/tal-tech/go-zero/core/stringx"
-	"github.com/tal-tech/go-zero/tools/goctl/model/sql/builderx"
+	"github.com/tal-tech/go-zero/core/stores/builderx"
 )
 `
 	// ImportsNoCache defines a import template for model in normal case
@@ -25,7 +25,7 @@ var (
 	"github.com/tal-tech/go-zero/core/stores/sqlc"
 	"github.com/tal-tech/go-zero/core/stores/sqlx"
 	"github.com/tal-tech/go-zero/core/stringx"
-	"github.com/tal-tech/go-zero/tools/goctl/model/sql/builderx"
+	"github.com/tal-tech/go-zero/core/stores/builderx"
 )
 `
 )
